requirements.txt
pydantic>=2.0
fastapi>=0.104
uvicorn>=0.24
httpx>=0.25
openai>=1.0
numpy>=1.24
python-dotenv>=1.0
exceptiongroup>=1.2
pytest>=9.0
pytest-asyncio>=1.0
